Move CLI dev server helpers to lib
